rename embeddings, move siren to mlp
