+++ neuralop/layers/embeddings.py
@@ -77,7 +77,7 @@ class RotaryEmbedding(nn.Module):
                           apply_rotary_pos_emb(t_y, freqs_y)), dim=-1)
 
 
-class GaussianFourierFeatureTransform(nn.Module):
+class GaussianFourierEmbedding(nn.Module):
     def __init__(self,
                  in_channels,
                  mapping_size,
@@ -121,123 +121,3 @@ class GaussianFourierFeatureTransform(nn.Module):
         x = 2 * torch.pi * x
 
         return torch.cat([torch.sin(x), torch.cos(x)], dim=-1)
-
-
-class Sine(nn.Module):
-    def __init__(self, w0=1.):
-        super().__init__()
-        self.w0 = w0
-
-    def forward(self, x):
-        return torch.sin(self.w0 * x)
-
-
-class Siren(nn.Module):
-    def __init__(self,
-                 dim_in,
-                 dim_out,
-                 w0=1.,
-                 c=6.,
-                 is_first=False,
-                 use_bias=True,
-                 activation=None):
-        """
-            code modified from:
-                 https://github.com/lucidrains/siren-pytorch/blob/master/siren_pytorch/siren_pytorch.py
-            SIREN paper: https://arxiv.org/abs/2006.09661
-            The Siren layer is a linear layer followed by a sine activation function.
-
-            Parameters:
-                dim_in: int, Number of input channels.
-                dim_out: int, Number of output channels.
-                w0: float, scaling factor (denominator) used to initialize the weights, by default 6.
-                c: float, scaling factor (numerator) used to initialize the weights, by default 6.
-                is_first: bool, Whether this is the first layer of the network, by default False.
-                use_bias: bool, Whether to use bias or not, by default True.
-                activation: nn.Module, Activation function to use, by default None (which uses Sine activation).
-
-        """
-
-        super().__init__()
-        self.dim_in = dim_in
-        self.is_first = is_first
-
-        weight = torch.zeros(dim_out, dim_in)
-        bias = torch.zeros(dim_out) if use_bias else None
-        self.init_(weight, bias, c=c, w0=w0)
-
-        self.weight = nn.Parameter(weight)
-        self.bias = nn.Parameter(bias) if use_bias else None
-        self.activation = Sine(w0) if activation is None else activation
-
-    def init_(self, weight, bias, c, w0):
-        dim = self.dim_in
-
-        w_std = (1 / dim) if self.is_first else (math.sqrt(c / dim) / w0)
-        weight.uniform_(-w_std, w_std)
-
-        if bias is not None:
-            bias.uniform_(-w_std, w_std)
-
-    def forward(self, x):
-        out = torch.nn.functional.linear(x, self.weight, self.bias)
-        out = self.activation(out)
-        return out
-
-
-class SirenNet(nn.Module):
-    def __init__(self,
-                 dim_in,
-                 dim_hidden,
-                 dim_out,
-                 num_layers,
-                 w0=1.,
-                 w0_initial=30.,
-                 use_bias=True,
-                 final_activation=None):
-        """
-            A MLP network with Siren layers.
-
-            Parameters:
-                dim_in: int, Number of input channels.
-                dim_hidden: int, Number of hidden channels.
-                dim_out: int, Number of output channels.
-                num_layers: int, Number of layers in the network.
-                w0: float, scaling factor (denominator) used to initialize the weights, by default 6.
-                w0_initial: float, scaling factor (denominator) used to initialize the weights for the first layer, by default 30.
-                use_bias: bool, Whether to use bias or not, by default True.
-                final_activation: nn.Module, Activation function to use in the final layer, by default None.
-        """
-        super().__init__()
-        self.num_layers = num_layers
-        self.dim_hidden = dim_hidden
-
-        self.layers = nn.ModuleList([])
-        for ind in range(num_layers):
-            is_first = ind == 0
-            layer_w0 = w0_initial if is_first else w0
-            layer_dim_in = dim_in if is_first else dim_hidden
-
-            self.layers.append(Siren(
-                dim_in=layer_dim_in,
-                dim_out=dim_hidden,
-                w0=layer_w0,
-                use_bias=use_bias,
-                is_first=is_first,
-            ))
-
-        self.final_activation = nn.Identity() if final_activation is None else final_activation
-        self.last_layer = Siren(dim_in=dim_hidden,
-                                dim_out=dim_out,
-                                w0=w0,
-                                use_bias=use_bias,
-                                activation=final_activation)
-
-    def forward(self, x):
-
-        for layer in self.layers:
-            x = layer(x)
-
-        x = self.last_layer(x)
-        x = self.final_activation(x)
-        return x
